requirements.txt
streamlit
numpy
plotly
requests
Pillow